requirements.txt
django>=2.2.5
django-allauth>=0.40.0
django-anymail[mailgun]>=7.0.0
celery>=4.3.0

